man: Ignore UNUSED attribute macros on parameters for doxygen docs
Since the UNUSED macros are placed after the parameter names, Doxygen parses the attribute (UNUSED) like if it was the parameter name, and the parameter name as if it was part of the type. Doxygen then can't find the correct parameter in the docstring, and complains that there's a parameter 'UNUSED' not documented, and that an extra parameter is documented but is missing in the function signature.

This also applies to the NOPG_UNUSED and NO_NLS_UNUSED macros.

In the Doxygen config, we can predefine macros with another definition, essentially removing them.
Note that this also means that permalinks built from a has of the signatures will change, but they were wrong before.